Expanded residual analysis, performance improvements
* Switch to single precision integers (for indices, etc.) -> big performance gain
* Include optional mask arguments in statistical functions (used to only
  analyze residual values above a specified threshold)
* Add median / stdev functionality to stats module, requires array module
  for sorting functionality (quicksort algorithm)

+++ numerics.f90
@@ -31,7 +31,7 @@ contains
     real(dp), intent(out) :: x_arr(:)
     real(dp), intent(out) :: dx
 
-    integer(dp) :: i_x, n_x
+    integer :: i_x, n_x
 
     n_x = size(x_arr)
     dx = (x_max - x_min) / n_x
@@ -55,7 +55,7 @@ contains
     complex(dp), intent(in) :: arr(:)
     complex(dp), intent(inout) :: d_arr(:)
     real(dp), intent(in) :: dx
-    integer(dp) :: i_x, n_x
+    integer :: i_x, n_x
     real(dp) :: scale
 
     n_x = size(arr)
@@ -92,7 +92,7 @@ contains
     complex(dp), intent(in) :: arr(:)
     complex(dp), intent(inout) :: d2_arr(:)
     real(dp), intent(in) :: dx
-    integer(dp) :: i_x, n_x
+    integer :: i_x, n_x
     real(dp) :: scale
 
     n_x = size(arr)
@@ -167,7 +167,7 @@ contains
     real(dp), intent(in) :: f_arr(:)
     real(dp), intent(in) :: dx
 
-    integer(dp) :: i_x
+    integer :: i_x
     real(dp) :: trapz_cnst
 
     trapz_cnst = dx / 2.0_dp
